pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comfyui-api-manager"
version = "2.0.0"
description = "Comprehensive API manager for ComfyUI with MCP server integration, workflow management, and job tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "ComfyUI API Manager Contributors" }
]
keywords = [
    "comfyui",
    "api",
    "mcp",
    "model-context-protocol",
    "workflow",
    "ai",
    "image-generation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.1",
    "websockets>=12.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "mcp>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/levidehaan/ComfyUI_API_Manager"
Repository = "https://github.com/levidehaan/ComfyUI_API_Manager"
Issues = "https://github.com/levidehaan/ComfyUI_API_Manager/issues"

[project.scripts]
comfyui-mcp-server = "comfyui_api_manager.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
